Fixed error of being able to delete tag or category associated with a post

+++ app/Http/Controllers/CategoriesController.php
@@ -105,6 +105,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category cannot be deleted becaue it has some posts');
+
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category successfully deleted');
